# api/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentação da API de delivery Cardapiogo.",
        "title": "Cardapiogo API",
        "contact": {
            "name": "Gustavo Ferreira de Jesus",
            "email": "gustavo081900@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/v1/store/item": {
            "get": {
                "description": "Returns a list of store items based on the provided filter criteria.\nSupports filtering by type, name, city, score, and maximum price.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Search and filter store items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by item type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by item name (partial match)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by store city",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by minimum score",
                        "name": "score",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by maximum price",
                        "name": "maxPrice",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of items matching the filter criteria",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/item.GetItemByIDOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter parameters",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "404": {
                        "description": "No items found",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Modifies the details of an existing store item. Only authenticated store owners can update their own items.\nAll fields in the update request will replace the existing values.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Update an existing store item",
                "parameters": [
                    {
                        "description": "Updated item details including name, description, price, and availability",
                        "name": "Params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/item.UpdateItemInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item successfully updated"
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Bearer token missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not the store owner",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict with existing data",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creates a new item in the store. Only authenticated store owners can perform this operation.\nThe item will be associated with the owner's store and available in the catalog.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Create a new store item",
                "parameters": [
                    {
                        "description": "Item creation parameters including name, description, price, and type",
                        "name": "Params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/item.CreateItemInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns the created item ID",
                        "schema": {
                            "$ref": "#/definitions/item.CreatedItem"
                        }
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Bearer token missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not the store owner",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict - Item already exists",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
        "/v1/store/item/{id}": {
            "get": {
                "description": "Returns detailed information about a specific store item by its ID.\nThis endpoint is publicly accessible and does not require authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Retrieve a specific store item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID - Unique identifier of the item",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item details including name, description, price, and availability",
                        "schema": {
                            "$ref": "#/definitions/item.GetItemByIDOutput"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Removes an item from the store catalog. Only authenticated store owners can delete their own items.\nThis operation is irreversible. All associated data will be permanently removed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item"
                ],
                "summary": "Delete a store item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item successfully deleted"
                    },
                    "400": {
                        "description": "Invalid item ID",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Bearer token missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User is not the store owner",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    },
                    "502": {
                        "description": "Bad gateway",
                        "schema": {
                            "$ref": "#/definitions/xerrors.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "item.CreateItemInput": {
            "type": "object",
            "required": [
                "description",
                "name",
                "price",
                "storeID",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "The best burguer!"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 25,
                    "example": "Hamburguer"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2990
                },
                "storeID": {
                    "type": "string",
                    "example": "3279gbf23gb-fb23-6f239"
                },
                "type": {
                    "enum": [
                        "FOOD",
                        " WATERGALLON"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/item.Type"
                        }
                    ]
                }
            }
        },
        "item.CreatedItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "item.GetItemByIDOutput": {
            "type": "object",
            "required": [
                "active",
                "description",
                "discountActive",
                "discountPrice",
                "image",
                "name",
                "price",
                "score",
                "storeID",
                "type"
            ],
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "description": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Com blend de 150g e molho especial."
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "discountActive": {
                    "type": "boolean",
                    "example": true
                },
                "discountPrice": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2499
                },
                "image": {
                    "type": "string",
                    "example": "https://cardapiogo.com.br/fgbjkgr7erg793bj3lk"
                },
                "name": {
                    "type": "string",
                    "maxLength": 25,
                    "example": "Hambúrguer Artesanal"
                },
                "price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 2990
                },
                "score": {
                    "type": "integer",
                    "example": 475
                },
                "storeID": {
                    "type": "string",
                    "example": "3279gbf23gb-fb23-6f239"
                },
                "type": {
                    "enum": [
                        "FOOD",
                        " WATERGALLON"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/item.Type"
                        }
                    ]
                }
            }
        },
        "item.Type": {
            "type": "string",
            "enum": [
                "FOOD",
                "WATERGALLON"
            ],
            "x-enum-varnames": [
                "Food",
                "WaterGallon"
            ]
        },
        "item.UpdateItemInput": {
            "type": "object",
            "required": [
                "active",
                "description",
                "discountActive",
                "discountPrice",
                "id",
                "name",
                "price",
                "storeID",
                "type"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string",
                    "maxLength": 50
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "discountActive": {
                    "type": "boolean"
                },
                "discountPrice": {
                    "type": "integer",
                    "minimum": 0
                },
                "id": {
                    "type": "integer",
                    "example": 246643
                },
                "name": {
                    "type": "string",
                    "maxLength": 25
                },
                "price": {
                    "type": "integer",
                    "minimum": 0
                },
                "storeID": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "FOOD",
                        " WATERGALLON"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/item.Type"
                        }
                    ]
                }
            }
        },
...
        "xerrors.CustomError": {
            "type": "object",
            "properties": {
                "debug": {},
                "details": {},
                "error": {
                    "type": "string"
                },
                "traceID": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}